refactor : test 패키지 변경 및 int number Wrapper 클래스로 변경
